requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
boto3==1.35.63
botocore==1.35.63
celery==5.4.0
celery-redbeat==2.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
DateTime==5.5
debugpy==1.8.11
dj-database-url==2.3.0
Django==5.1.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-environ==0.11.2
django-heroku==0.3.1
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
frozendict==2.4.6
frozenlist==1.5.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
Jinja2==3.1.5
jmespath==1.0.1
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
mccabe==0.7.0
multidict==6.1.0
multitasking==0.0.11
mysqlclient==2.2.6
nulltype==2.3.1
numpy==2.2.2
packaging==24.2
pandas==2.2.3
peewee==3.17.8
phonenumbers==8.13.52
plaid-python==28.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pycurl==7.45.4
pyflakes==3.2.0
PyJWT==2.9.0
pyotp==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
s3transfer==0.10.3
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tenacity==9.0.0
timedelta==2020.12.3
twilio==9.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.44.0
whitenoise==6.8.2
yarl==1.18.3
yfinance==0.2.52
zope.interface==7.1.1
